pyproject.toml
[project]
name = "knowledge-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno[os,telegram]>=2.5.17",
    "openai>=2.32.0",
    "sentence-transformers",
    "huggingface-hub",
    "chromadb",
    "python-dotenv>=1.2.2",
    "docling>=2.91.0",
    "openrouter>=0.9.1",
    "mcp>=1.27.0",
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
    "requests>=2.33.1",
]
